requirements.txt
h5py==3.1.0
matplotlib==3.9.4
NiBabel==4.0.2
nilearn==0.9.2
numpy==1.26.4
pandas==1.5.1
pillow==9.2.0
pycortex==1.2.10
scikit-learn==1.1.1
scipy==1.12.0
torch==1.13.0
tqdm==4.64.1
torchvision==0.14.0
